pyproject.toml
[project]
requires-python = ">=3.10,<3.14"
name = "torch_cuda_uv_demo"
version = "0.1.0"
dependencies = [
    "torch>=2.7.0,<3.0",
    "torchvision>=0.22.0,<1.0",
    "torchaudio>=2.7.0,<3.0",
    "opencv-python>=4.12.0.88",
    "timm>=1.0.22",
    "onnxscript>=0.5.7",
    "onnxruntime-gpu>=1.23.2",
]

# ① 声明 PyTorch 的 CUDA 12.6 索引
[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

# ② 告诉 uv：torch/vision/audio 都从这个 CUDA 索引装（Win + Linux）
[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
